fixed app file
