+++ server/tests/test_chatbot.py
@@ -20,7 +20,7 @@ if parent_dir not in sys.path:
 # Import the preprocess_input function from src.chatbot
 from src.chatbot import preprocess_input
 
-from src.chatbot import match_intent_to_response, read_data
+from src.chatbot import match_intent_to_response
 
 def test_preprocess_normal_input():
     input_text = "This is a sample message with punctuation! Very nice."
@@ -56,19 +56,3 @@ def test_match_intent_to_response():
                       } 
     assert match_intent_to_response(intent, dictionary) == expected_output
 
-def test_read_data():
-    with patch('model.pd.read_excel') as mock_read_excel:
-        mocked_data = pd.DataFrame([
-            ["Forgot_credentials", "Please specify which credential you need: Password, Username, or Email"],
-            ["Password_reset", "Please click on the following link to reset your password: https://myaccount.ucdenver.edu/change-password"],
-            ["Username", "To obtain your username, you can go to the following website and click on 'Forgot my username?' under the 'Next Step' button. Your username will be sent to the email you registered your account with: https://myaccount.ucdenver.edu/change-password"]
-        ])
-        mock_read_excel.return_value = mocked_data
-        dictionary = {
-            'Forgot_credentials': "Please specify which credential you need: Password, Username, or Email",
-            "Password_reset": "Please click on the following link to reset your password: https://myaccount.ucdenver.edu/change-password",
-            "Username": "To obtain your username, you can go to the following website and click on 'Forgot my username?' under the 'Next Step' button. Your username will be sent to the email you registered your account with: https://myaccount.ucdenver.edu/change-password"
-        }
-
-        result = read_data()
-        assert result == dictionary
